requirements.txt
httpx==0.27.0
numpy==1.24.3
pandas==2.0.3
torch==2.2.2
sentence-transformers==3.0.1
python-dotenv==1.0.0
python-telegram-bot==21.3
vaderSentiment==3.3.2
transformers==4.43.1
apscheduler==3.10.4
langchain==0.2.13
langchain-openai==0.1.21
openai>=1.40.0
croniter==1.3.15
cffi==1.14.6
pytz==2024.1
boto3
